cat cat

+++ Week01/STR/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: agrele <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/24 18:27:58 by agrele            #+#    #+#             */
-/*   Updated: 2015/11/24 18:44:31 by agrele           ###   ########.fr       */
+/*   Updated: 2015/11/26 14:47:38 by agrele           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-int		ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int len;
+	size_t len;
 
 	len = 0;
 	while (s[len])
